% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_format_guess}
\alias{file_format_guess}
\title{根据操作系统参数猜后缀名}
\usage{
file_format_guess(osName = "ubuntu")
}
\arguments{
\item{osName}{操作系统名单}
}
\value{
返回后缀名
}
\description{
根据操作系统参数猜后缀名
}
\examples{
file_format_guess('ubuntu')
}
