% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{fileName}
\alias{fileName}
\title{形成文件名}
\usage{
fileName(appName = "shinyserver", appVersion = "lastest",
  osName = "ubuntu")
}
\arguments{
\item{appName}{主文件名称}

\item{appVersion}{主文件版本}

\item{osName}{操作系统名称}
}
\value{
返回值
}
\description{
形成文件名
}
\examples{
fileName('shinyserver','lastest','ubuntu')

}
