% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc}
\alias{jdbc}
\title{下载jdbc驱动}
\usage{
jdbc(dbName = "sqlserver", dbVersion = "lastest",
  repo = "./data-raw")
}
\arguments{
\item{dbName}{数据库名称}

\item{dbVersion}{数据库版本}

\item{repo}{下载存放目录}
}
\value{
返回值
}
\description{
下载jdbc驱动
}
\examples{
jdbc()
library(tsfs)
jdbc()
jdbc('sqlserver','7.2.2')
jdbc('oracle')
}
