% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_fileName}
\alias{jdbc_fileName}
\title{设置数据库jdbc文件名}
\usage{
jdbc_fileName(dbName = "sqlserver", dbVersion = "lastest")
}
\arguments{
\item{dbName}{数据库名称}

\item{dbVersion}{版本}
}
\value{
返回值
}
\description{
设置数据库jdbc文件名
}
\examples{
jdbc_fileName()
}
