% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_repo}
\alias{jdbc_repo}
\title{创建下载的的目录}
\usage{
jdbc_repo(dbName = "sqlserver", dbVersion = "lastest",
  repo = "./data-raw")
}
\arguments{
\item{dbName}{数据库名称}

\item{dbVersion}{版本}

\item{repo}{目录}
}
\value{
返回值
}
\description{
创建下载的的目录
}
\examples{
jdbc_repo()
}
