% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_url}
\alias{jdbc_url}
\title{设置下载链接}
\usage{
jdbc_url(dbName = "sqlserver", dbVersion = "lastest")
}
\arguments{
\item{dbName}{数据库名称}

\item{dbVersion}{数据库版本}
}
\value{
返回值
}
\description{
设置下载链接
}
\examples{
jdbc_url()
}
