% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{pull}
\alias{pull}
\title{提取数据}
\usage{
pull(appName = "shinyserver", appVersion = "lastest",
  osName = "ubuntu", repo = "./")
}
\arguments{
\item{appName}{文件名称}

\item{appVersion}{文件版本}

\item{osName}{操作系统名称}

\item{repo}{下载路径}
}
\value{
下载成功返回TRUE否则返回FALSE
}
\description{
提取数据
}
\examples{
library(tsfs)
bb <- pull(appName = 'rstudioserver',appVersion = '1.2.5033',osName = 'centos6',repo = './data-raw')
print(bb)
bb <- pull(appName = 'rstudioserver',appVersion = '1.2',osName = 'centos6',repo = './data-raw')
print(bb)
bb <- pull(appName = 'rstudioserver',appVersion = 'lastest',osName = 'centos6',repo = './data-raw')
print(bb)
bb <- pull(appName = 'rstudioserver',appVersion = '1.2.5033',osName = 'centos7',repo = './data-raw')
print(bb)
bb <- pull(appName = 'rstudioserver',appVersion = '1.2',osName = 'centos7',repo = './data-raw')
print(bb)
bb <- pull(appName = 'rstudioserver',appVersion = 'lastest',osName = 'centos7',repo = './data-raw')
print(bb)
bb <- pull(appName = 'rstudioserver',appVersion = '1.2.5033',osName = 'centos8',repo = './data-raw')
print(bb)
bb <- pull(appName = 'rstudioserver',appVersion = '1.2',osName = 'centos8',repo = './data-raw')
print(bb)
bb <- pull(appName = 'rstudioserver',appVersion = 'lastest',osName = 'centos8',repo = './data-raw')
print(bb)
}
