% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{py_pkgs}
\alias{py_pkgs}
\title{安装python包}
\usage{
py_pkgs(pkg_name = "pandas", version = "1.0.3", os = "linux64",
  py_ver = "py37", repo = "./data-raw")
}
\arguments{
\item{pkg_name}{包名称}

\item{version}{包版本}

\item{os}{操作系统}

\item{py_ver}{python版本}

\item{repo}{存放目录}
}
\value{
返回值
}
\description{
安装python包
}
\examples{
py_pkgs()
library(tsfs)
tsfs::py_pkgs()
tsfs::py_pkgs(os='linux32')
tsfs::py_pkgs(pkg_name = 'numpy',version = '1.18.2',os = 'linux64',py_ver = 'py37')
}
