% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pypkg_fileName}
\alias{pypkg_fileName}
\title{pypkgs文件名}
\usage{
pypkg_fileName(pkg_name = "pandas", version = "1.0.3",
  os = "linux64", py_ver = "py37")
}
\arguments{
\item{pkg_name}{包名称}

\item{version}{版本}

\item{os}{操作系统}

\item{py_ver}{python版本}
}
\value{
返回链接
}
\description{
pypkgs文件名
}
\examples{
pypkg_fileName()
}
