% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pypkg_repo}
\alias{pypkg_repo}
\title{形成基地路径}
\usage{
pypkg_repo(pkg_name = "pandas", version = "1.0.3", os = "linux64",
  py_ver = "py37", repo = "./data-raw")
}
\arguments{
\item{pkg_name}{包名}

\item{version}{版本}

\item{os}{操作系统}

\item{py_ver}{python版本}

\item{repo}{存储路径}
}
\value{
返回完整路径
}
\description{
形成基地路径
}
\examples{
pypkg_repo()
}
