% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{repo_fileName}
\alias{repo_fileName}
\title{形成本地的文件}
\usage{
repo_fileName(appName = "shinyserver", appVersion = "lastest",
  osName = "ubuntu", repo = "./")
}
\arguments{
\item{appName}{文件名}

\item{appVersion}{文件版本}

\item{osName}{操作系统}

\item{repo}{下载路径}
}
\value{
返回值
}
\description{
形成本地的文件
}
\examples{
repo_fileName()
}
