% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{rstudio_server}
\alias{rstudio_server}
\title{下载rstudio server文件}
\usage{
rstudio_server(os = "ubuntu18", repo = "./data-raw",
  version = "lastest")
}
\arguments{
\item{os}{操作系统名称}

\item{repo}{下载目录}

\item{version}{版本号}
}
\value{
返回值
}
\description{
下载rstudio server文件
}
\examples{
rstudio_server()
library(tsfs)
tsfs::rstudio_server('ubuntu16')
tsfs::rstudio_server('ubuntu18')
tsfs::rstudio_server('centos7')
}
